Added Tatome's version of the Hartigans' dip test for unimodality.  Not sure how well it works in this type of data context, but it's available now.  The library is not packaged for any kind of install, so I just dropped in the code and added his license to this repository.
Added a script to filter seriation data sets by a variety of criteria, including the dip test, removal of all zero columns,
and removal of any columns that have less than 2-3 entries.  The idea here is to replicate the series of things that culture
historians tended to do intuitively when building chronologies using seriation, including the "test of historical significance."

moved explorations to notebooks.

Created annotation as library module, for use in a variety of scripts.